modelIA/requirements.txt
flask
flask-cors
tensorflow
pillow
gunicorn
pytest
